Port the `Client` class
